Dataframe the documentation of installed packages.

This is fundamentally a dataframe version of the output of utils::hsearch_db().

Usage

search_docs(packages = NULL)

Arguments

packages

A character vector with the names of packages to limit the search, or NULL to serach in all known libraries.

Value

A dataframe.

See also

Examples

search_docs(NULL)
#> # A tibble: 39,339 × 11
#>    package  libpath id    name  title topic encoding type  alias keyword concept
#>    <chr>    <chr>   <chr> <chr> <chr> <chr> <chr>    <chr> <chr> <chr>   <chr>  
#>  1 R6       /home/… 1/1   R6Cl… Crea… R6Cl… UTF-8    help  R6Cl… NA      NA     
#>  2 R6       /home/… 1/1   R6Cl… Crea… R6Cl… UTF-8    help  R6    NA      NA     
#>  3 R6       /home/… 1/2   as.l… Crea… as.l… UTF-8    help  as.l… NA      NA     
#>  4 R6       /home/… 1/3   is.R6 Is a… is.R6 UTF-8    help  is.R6 NA      NA     
#>  5 R6       /home/… 1/3   is.R6 Is a… is.R6 UTF-8    help  is.R… NA      NA     
#>  6 RColorB… /home/… 2/1   RCol… Colo… RCol… UTF-8    help  brew… color   Color,…
#>  7 RColorB… /home/… 2/1   RCol… Colo… RCol… UTF-8    help  disp… color   Color,…
#>  8 RColorB… /home/… 2/1   RCol… Colo… RCol… UTF-8    help  disp… color   Color,…
#>  9 RColorB… /home/… 2/1   RCol… Colo… RCol… UTF-8    help  brew… color   Color,…
#> 10 RColorB… /home/… 2/1   RCol… Colo… RCol… UTF-8    help  brew… color   Color,…
#> # ℹ 39,329 more rows