Dataframe the documentation of installed packages.

This is fundamentally a dataframe version of the output of utils::hsearch_db().

search_docs(packages = NULL)

Arguments

packages
A character vector with the names of packages to limit the search, or NULL to serach in all known libraries.

Value

A dataframe.

See also

Examples

search_docs(NULL)

#> # A tibble: 34,336 x 11
#>    package  libpath id    name  title topic encoding type  alias keyword concept
#>    <chr>    <chr>   <chr> <chr> <chr> <chr> <chr>    <chr> <chr> <chr>   <chr>  
#>  1 askpass  /home/… 1/1   askp… Pass… askp… UTF-8    help  askp… <NA>    <NA>   
#>  2 askpass  /home/… 1/2   ssh_… ASKP… ssh_… UTF-8    help  ssh_… <NA>    <NA>   
#>  3 assertt… /home/… 2/1   are_… Are … are_… ""       help  are_… <NA>    <NA>   
#>  4 assertt… /home/… 2/2   asse… Miss… is.e… ""       help  is.e… <NA>    <NA>   
#>  5 assertt… /home/… 2/2   asse… Miss… is.e… ""       help  is.t… <NA>    <NA>   
#>  6 assertt… /home/… 2/2   asse… Miss… is.e… ""       help  is.d… <NA>    <NA>   
#>  7 assertt… /home/… 2/3   asse… Asse… asse… ""       help  asse… <NA>    <NA>   
#>  8 assertt… /home/… 2/3   asse… Asse… asse… ""       help  see_… <NA>    <NA>   
#>  9 assertt… /home/… 2/4   asse… Usef… asse… ""       help  asse… <NA>    <NA>   
#> 10 assertt… /home/… 2/4   asse… Usef… asse… ""       help  is.d… <NA>    <NA>   
#> # … with 34,326 more rows